Add ActiveRecord back into app
This is mostly a revert of 0ce5a26752728f28b84ec49e9543b6ec9de699fc, but
I did it by hand so that I could make some choices along the way. I
prefer the SQL schema file, for example and of course chose Postgres
over the default sqlite. :wink:

I also went back to the setup script and added what would be required
there so that a fresh install would have the database setup properly.

That's all I can think of, but maybe there are other places to update?