Organize Procfiles
I like to have the dev version of this file use a standard Rails server
command so that I see more in the logs. I also moved the webpack item to
the dev version - I don't think this made sense to send to Heroku, but
maybe I'm missing something there.