Fix overeager pre-population of the common genome.
Before this commit I was observing that the selection of gene inputs
was occasionally getting wiped away and replaced with the common genome
at unexpected times. That's because state changes higher up in the app were
causing downstream components to update (as expected) but we were not checking
to see if we actually *needed* to initialize the form to the common genome,
we were just doing it recklessly. Fixed now!