Prevent double-scrollbar when modal is open.
Before this commit, when the modal was open we might inadverently
be able to scroll the underlying body content, and see scrollbars
in passing, accordingly.

In fact, if the modal *itself* had long i.e. scrollable content, then
we'd see a double scrollbar.

This prevents all that by temporarily adding a "overflow: hidden"
css rule to the <body> tag, while the modal is open.