Add a spacer so that gene autocomplete has room to show suggestions.
This builds in a little whitespace below the gene autocomplete in order
to avoid an annoying UI quirk — having to scroll twice: once to get down
to the autocomplete widget, and once again to scroll the autocomplete suggestions
into view.

Because the suggestions don't exist in the DOM at the time you initially scroll down
we now build in some whitespace to "reserve" that area so that you can scroll down
sufficiently far in one move.