Fix user_id column type
I made a mistake and should have used the string type for the Gravity
user id. Since it was an integer, what was happening was that Rosalind
was ending up with only the first N number characters of the Gravity
user id. :\

This migration fixes that mistake, but then leaves us with those partial
ids. Once this migration has been run on production, then the next step
is to convert the partial ids to the full ones:

```ruby
map = { 'partial' => 'full' }
map.each { |key, value| BatchUpdate.where(user_id: key).update_all(user_id: value) }
```

I worked with @anandaroop and @reganartsy on coming up with the map that
will fix this so I'm ready once this lands on production.