Fix production Heroku deployments
Totally mixed this up here - I had this if/else structure in the deploy
file, but then wasn't actually using it anywhere! What makes more sense
is to move this to the setup script and get the right remote to begin
with and then you don't have to specify app-name for the heroku commands
at all. This only revealed itself as an issue when I went to deploy to
production.