Convert to headless Chrome testing
Since headless mode landed in the Chrome stable channel earlier
this year, we've had the option of running our Capybara tests
headlessly.

This implements that, with some cues from
https://robots.thoughtbot.com/headless-feature-specs-with-chrome

We leave both :headless_chrome and :chrome as driver options
so that devs can temporarily drop into headful mode for a test run
if they need to.