Use before hook
I was noticing some test suite flickering, that I was able to chase down
like this:

```
rspec --seed 36230 --bisect
```

Which gave me this:

```
rspec './spec/features/user_updates_artworks_spec.rb[1:2]' './spec/jobs/process_batch_update_job_spec.rb[1:2:1]' --seed 36230
```

Note: it turns out either of the tests in
`process_batch_update_job_spec` had this issue, it was just a
coincidence that the second one was in the flicker I saw.

So, after some tinkering, I noticed that if I moved the clearing of jobs
to the before hook, then the flickering went away. There might be a bug
in the framework here somewhere, but I can't really figure out how to
troubleshoot anymore than this, so I'm calling it good.