Use Webpacker's version of Webpack
This gets rid of a weird bug where the Webpack-generated manifest.json
file was coming up empty.

Took me a bit of staring back and forth betweeen this repo and a
pristine 5.2 Webpacker one to see that we didn't need a "webpack"
entry in our package.json — it's a transitive dependency of
"@rails/webpacker" so let's just rely on that one instead.